Expose SchemaManifest publicly
- Move SchemaManifest, FromParquetSchema* functions into
  parquet/arrow/shcema.h
- Rename BuildSchemaManifest to SchemaManifest::Make to follow the code
  implicit standard
- Refactor RowGroupStatisticsAsExpression to use SchemaManifest
